requirements.txt
requests>=2.31.0
feedparser>=6.0.10
beautifulsoup4>=4.12.0
lxml>=4.9.0
colorama>=0.4.6
rich>=13.0.0
python-dateutil>=2.8.0
aiohttp>=3.8.0
asyncio>=3.4.3
python-dotenv>=1.0.0